Decouple StreamProcessor construction and decoder configuration (#3223)
Summary:
Pull Request resolved: https://github.com/pytorch/audio/pull/3223

Each `StreamProcessor` is responsible for processing a source stream. In the case where we support packet passthrough, `StreamProcessor`'s choice of decoder is irrelevant as no decoding is performed. Currently, however, `StreamProcessor` requires decoder params and fixes a decoder at construction time. To accommodate this future packet passthrough use case, this PR decouples the construction of `StreamProcessor` from the configuration of the decoder that it uses.

Reviewed By: mthrok

Differential Revision: D44554934

fbshipit-source-id: 1d1a89015e1181b71dfb95c928de4fc3ec6f63b6