[BC-breaking] Make io optional arguments kw-only
Recently, we added bunch of options to make StreamReader/Writer
flexible. As a result, their methods have many number of arguments,
and some of them have semantic grouping.

For example, the arguments of ``StreamWriter.add_video_stream``
are roughly grouped as follow;

- Information about input media format
  frame_rate, width, height, format
- Information about encoder
  encoder, encoder_option
- Information about codec configuration
  codec_config
- Information about encode media format
  encoder_format, encoder_frame_rate, encoder_width, encoder_height
- Information about additional processing
  filter_desc
- Hardware acceleration
  hw_accel

We do not know what arguments add in the future, and we want to keep
them roughly grouped, by inserting a new argment in a middle.

This commit puts most of them in keyword-only argument, so that
we can rearrange them without breaking backward compatibility.