Remove linux GPU unittest from CircleCI
Linux GPU unit test on CircleCI relies on custom Docker image
with CUDA 10.2.

PyTorch 2.0 does not support CUDA 10, so these tests have not run
for a while.

We have GPU tests on GHA for Linux, so we can get rid of them.

Windows GPU tests are not ported to GHA yet, but they are still
working on CircleCI, so we don't delete them yet.