Remove linux GPU unit test from CircleCI (#3231)
Summary:
Linux GPU unit test on CircleCI relies on custom Docker image with CUDA 10.2.

PyTorch 2.0 does not support CUDA 10, so these tests have not run for a while.

We have GPU tests on GHA for Linux, so we can get rid of them.

Windows GPU tests are not ported to GHA yet, but they are still working on CircleCI, so we don't delete them yet.

Pull Request resolved: https://github.com/pytorch/audio/pull/3231

Reviewed By: hwangjeff

Differential Revision: D44639302

Pulled By: mthrok

fbshipit-source-id: c1fd39f4805a50a12af4259d423985fe453fd229