Update and deprecate apply_codec function
To prepare for the upcoming removal of file-like object support from sox_io backend,
this commit changes apply_codec function to use tempfile.

`apply_codec` function is now deprecated and users are encourated to use
`torchaudio.io.AudioEffector`.

We will not remove the function itself, but will remove the entry from the doc.