Delete CCI Linux and MacOS Unittest Jobs (#3391)
Summary:
Deprecates the Linux and MacOS Unittest jobs now that they've been running on Nova for over a week.

Aside: There was also a stylecheck job that was dependent on the Linux Unittest job. I also put up https://github.com/pytorch/audio/pull/3390 to move that stylecheck job to Nova. I'm happy to reintroduce the CCI stylecheck job standalone in CCI if we want the Nova version to run on main for a week.

Pull Request resolved: https://github.com/pytorch/audio/pull/3391

Reviewed By: mthrok

Differential Revision: D46324198

Pulled By: osalpekar

fbshipit-source-id: 2115748e153c5dee1a38db2b6230acebc4f56927