Update the logic to fetch pixel format from filter graph
When using GPU decoder in some environments, attempting to read the
output formats from filter graph caused an issue in which the software
pixel format cannot be determined.

We do not know the exact cause but when it happens, the input link of
buffer sink does not have HW frames context.

Since currently no filter can convert the pixel format of CUDA frame,
we resort to the HW frames context of the output link of buffer source.

- V100,       CUDA 11.4, libavfilter 8.44.100, ffnvcodec 11.1.5.2
- RTX3080 Ti, CUDA 12.1, libavfilter 8.44.100, ffnvcodec 12.0.16.0