Remove ffmpeg fallback from sox_io backend
In #2419, we added ffmpeg as fallback for sox_io backend.
The was a warkaround for solving the issue with libmad removal.

Now that we introduced `backend` argument to I/O functions,
and libsox integration is moved to dynamic binding where users can
use libsox with libmad integration, we do not need the workaround.

This commit is based on reverting #2416 (fd7ace17938c74d0928987b9525dbe7799b328fa).