Fix style check CI job
It seems that the default Python version was updated to 3.11.
libcst does not have binary release for 3.11, so the CI attempts to
build from source but it fails because building libcst requires Rust
compiler.

This commit fix the Python version of style check job to 3.10 so that
the issue with Rust compiler is avoided.