Fix style check CI job (#3564)
Summary:
It seems that the default Python version was updated to 3.11.
libcst does not have binary release for 3.11, so the CI attempts to
build from source but it fails because building libcst requires Rust
compiler.

This commit fix the Python version of style check job to 3.10 so that
the issue with Rust compiler is avoided.

Pull Request resolved: https://github.com/pytorch/audio/pull/3564

Reviewed By: huangruizhe

Differential Revision: D48499560

Pulled By: mthrok

fbshipit-source-id: 53ab77268d8143f4946d92e8cd1f96aea55e7b72