retain arguments from init as properties in MelScale and InverseMelScale
MelScale should retain n_stft as an attribute
Fixes #4122

I observed inconsistency between the MelScale and InverseMelScale in which arguments were retained as attributes. Perhaps there was some reason for this, but I think it probably makes sense to save all arguments as attributes for both classes. This allows serialization of the object for re-initialization with the same parameters.